[DONE] Finish building ModelManager
