+++ src/agentscope/manager/__init__.py
@@ -2,7 +2,9 @@
 """Import all manager related classes and functions."""
 
 from ._file import FileManager
+from ._model import ModelManager
 
 __all__ = [
     "FileManager",
+    "ModelManager",
 ]
