+++ tests/litellm_test.py
@@ -4,8 +4,7 @@ import unittest
 from unittest.mock import patch, MagicMock
 
 import agentscope
-from agentscope.manager import FileManager
-from agentscope.models import load_model_by_config_name
+from agentscope.manager import FileManager, ModelManager
 
 
 class TestLiteLLMChatWrapper(unittest.TestCase):
@@ -48,7 +47,9 @@ class TestLiteLLMChatWrapper(unittest.TestCase):
             disable_saving=True,
         )
 
-        model = load_model_by_config_name("test_config")
+        model = ModelManager.get_instance().get_model_by_config_name(
+            "test_config",
+        )
 
         response = model(
             messages=self.messages,
